+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Game/Subsystems/Extension.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Game;
+package org.firstinspires.ftc.teamcode.Game.Subsystems;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 import com.qualcomm.robotcore.hardware.CRServo;
 public class Extension {
