+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Game/Subsystems/Intake.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.Game;
+package org.firstinspires.ftc.teamcode.Game.Subsystems;
 
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.HardwareMap;
