+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/Game/Subsystems/Spindex.java
@@ -1,9 +1,8 @@
-package org.firstinspires.ftc.teamcode.Game;
+package org.firstinspires.ftc.teamcode.Game.Subsystems;
 
 import com.qualcomm.robotcore.hardware.AnalogInput;
 import com.qualcomm.robotcore.hardware.CRServo;
 import com.qualcomm.robotcore.hardware.HardwareMap;
-import com.qualcomm.robotcore.hardware.Servo;
 
 public class Spindex {
     private CRServo spindex = null;
